requirements.txt
fastapi
uvicorn
aiogram==3.3.0
requests
python-dotenv
pydantic